third_party/WebKit/LayoutTests/fast/forms/setCustomValidity-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test how setCustomValidity reacts to too few arguments.');
var b = document.createElement('button');
shouldThrow("b.setCustomValidity()");
var fs = document.createElement('fieldset');
shouldThrow("fs.setCustomValidity()");
var i = document.createElement('input');
shouldThrow("i.setCustomValidity()");
</script>
</body>
</html>
